requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.5
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
anyio==4.3.0
    # via httpx
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   httpx
    #   yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
yarl==1.9.4
    # via aiohttp
